+++ src/HOC/modules/EconomicEventManager/EconomicEventManager.graphql
@@ -9,3 +9,27 @@ spatialThingsPages {
   }
 }
 
+query economicEventsFiltered($action: ID!) {
+economicEventsFiltered(action: $action){
+
+    id
+    provider {
+      id
+      name
+    }
+    receiver {
+      id
+      name
+    }
+  }
+}
+
+query unitsPages {unitsPages {
+    edges {
+      id
+    symbol
+    label
+    }
+}
+}
+
