+++ src/HOC/modules/EconomicEventManager/EconomicEventManagerHOC.tsx
@@ -5,25 +5,42 @@ import { useActionsQuery } from '../IntentPanel/Actions.generated';
 import EconomicEventManager, {
   EconomicEventManagerProps
 } from '../../../ui/modules/EconomicEventManager';
-import * as GQL from './EconomicEventManager.generated';
-import { useUnitsPagesQuery } from './EconomicEventUnits.generated';
+import {
+  useEconomicEventsFilteredQuery,
+  useUnitsPagesQuery
+} from './EconomicEventManager.generated';
 
 export const EconomicEventManagerHOC: FC = () => {
+  const [providerList, setProviderList] = React.useState<
+    null | undefined | [] | { id: string; name: string }[]
+  >([]);
+  const [receiverList, setReceiverList] = React.useState<
+    null | undefined | [] | { id: string; name: string }[]
+  >([]);
   const [action, setAction] = React.useState('');
 
   const intentActionsQ = useActionsQuery();
-  const { loading, error, data } = GQL.useSpatialThingsPagesQuery({
+  const { data } = useEconomicEventsFilteredQuery({
     variables: { action }
   });
+
+  React.useEffect(() => {
+    const providers =
+      data && data.economicEventsFiltered && data?.economicEventsFiltered.map(el => el.provider);
+    const receivers =
+      data && data.economicEventsFiltered && data?.economicEventsFiltered.map(el => el.receiver);
+    providers?.length && setProviderList(providers || null);
+    receivers?.length && setReceiverList(receivers || []);
+  }, [data]);
 
   const unitPagesQ = useUnitsPagesQuery();
 
   const intentActions = intentActionsQ.data?.actions;
   const unitPages = unitPagesQ.data?.unitsPages;
-
   const economicEventManager: EconomicEventManagerProps = {
     actionList: intentActions,
-    economicEvents: { loading, error, data: data },
+    providerList: providerList,
+    receiverList: receiverList,
     unitPages,
     setAction
   };
