--- src/HOC/modules/EconomicEventManager/EconomicEventUnits.graphql
@@ -1,9 +1,0 @@
-query unitsPages {unitsPages {
-    edges {
-      id
-    symbol
-    label
-    }
-}
-}
-
