--- src/HOC/modules/EconomicEventManager/EconomicEvenеtFiltered.graphql
@@ -1,14 +1,0 @@
-query economicEventsFiltered($action: ID!) {
-economicEventsFiltered(action: $action){
-
-    id
-    provider {
-      id
-      name
-    }
-    receiver {
-      id
-      name
-    }
-  }
-}
