+++ src/HOC/modules/previews/thread/ThreadPreview.generated.tsx
@@ -4,12 +4,13 @@ import { FlagPreviewFragment } from '../flag/FlagPreview.generated';
 import { ResourcePreviewFragment } from '../resource/ResourcePreview.generated';
 import { CollectionPreviewFragment } from '../collection/CollectionPreview.generated';
 import { CommunityPreviewFragment } from '../community/CommunityPreview.generated';
+import { CommentPreviewFragment } from '../comment/CommentPreview.generated';
 import gql from 'graphql-tag';
 import { CommunityPreviewFragmentDoc } from '../community/CommunityPreview.generated';
 import { CollectionPreviewFragmentDoc } from '../collection/CollectionPreview.generated';
 import { ResourcePreviewFragmentDoc } from '../resource/ResourcePreview.generated';
 import { FlagPreviewFragmentDoc } from '../flag/FlagPreview.generated';
-import { CommentPreviewFragment, CommentPreviewFragmentDoc } from '../comment/CommentPreview.generated';
+import { CommentPreviewFragmentDoc  } from '../comment/CommentPreview.generated';
 
 
 
@@ -19,26 +20,26 @@ export type ThreadPreviewFragment = (
   { __typename: 'Thread' }
   & Pick<Types.Thread, 'id' | 'lastActivity' | 'createdAt'>
   & { context: Types.Maybe<{ __typename: 'Category' } | (
-    { __typename: 'Collection' }
-    & CollectionPreviewFragment
-  ) | { __typename: 'Comment' } | (
-    { __typename: 'Community' }
-    & CommunityPreviewFragment
-  ) | (
-    { __typename: 'Flag' }
-    & FlagPreviewFragment
-  ) | { __typename: 'Follow' } | { __typename: 'Intent' } | { __typename: 'Like' } | { __typename: 'Organisation' } | (
-    { __typename: 'Resource' }
-    & ResourcePreviewFragment
-  ) | { __typename: 'SpatialThing' } | { __typename: 'Taggable' } | { __typename: 'User' }>, comments: Types.Maybe<(
-    { __typename: 'CommentsPage' }
-    & Pick<Types.CommentsPage, 'totalCount'>
-    & { edges: Array<(
-      { __typename: 'Comment' }
-      & CommentPreviewFragment
-    )> }
-  )> }
-);
+      { __typename: 'Collection' }
+      & CollectionPreviewFragment
+      ) | { __typename: 'Comment' } | (
+      { __typename: 'Community' }
+      & CommunityPreviewFragment
+      ) | (
+      { __typename: 'Flag' }
+      & FlagPreviewFragment
+      ) | { __typename: 'Follow' } | { __typename: 'Intent' } | { __typename: 'Like' } | { __typename: 'Organisation' } | (
+      { __typename: 'Resource' }
+      & ResourcePreviewFragment
+      ) | { __typename: 'SpatialThing' } | { __typename: 'Taggable' } | { __typename: 'User' }>, comments: Types.Maybe<(
+      { __typename: 'CommentsPage' }
+      & Pick<Types.CommentsPage, 'totalCount'>
+      & { edges: Array<(
+          { __typename: 'Comment' }
+          & CommentPreviewFragment
+          )> }
+      )> }
+  );
 
 export const ThreadPreviewFragmentDoc = gql`
     fragment ThreadPreview on Thread {
@@ -62,12 +63,15 @@ export const ThreadPreviewFragmentDoc = gql`
   comments(limit: 1) {
     totalCount
     edges {
-      ...CommentPreview
+      ... on Comment {
+        ...CommentPreview
+      }
     }
   }
 }
-    ${CommunityPreviewFragmentDoc}
+${CommunityPreviewFragmentDoc}
 ${CollectionPreviewFragmentDoc}
 ${ResourcePreviewFragmentDoc}
 ${FlagPreviewFragmentDoc}
-${CommentPreviewFragmentDoc}`;
+${CommentPreviewFragmentDoc}
+`;
