+++ src/ui/modules/EconomicEventManager/index.tsx
@@ -22,28 +22,10 @@ export type EconomicEventVariables = {
   hasNumericalValue: number | string | undefined;
 };
 
-type Person = {
-  id: string;
-  name: string;
-  __typename: string;
-};
-
-type UnitPage = {
-  id: string;
-  label: string;
-  symbol: string;
-};
-
-type EconomicEvent = {
-  id: string;
-  provider: Person;
-  receiver: Person;
-  __typename: string;
-};
-
 export type EconomicEventManagerProps = {
   actionList?: any;
-  economicEvents?: any;
+  providerList: null | undefined | [] | { id: string; name: string }[];
+  receiverList: null | undefined | [] | { id: string; name: string }[];
   unitPages?: {
     edges: any;
   };
@@ -53,7 +35,8 @@ export type EconomicEventManagerProps = {
 export const EconomicEventManager: React.FC<EconomicEventManagerProps> = ({
   actionList,
   setAction,
-  economicEvents,
+  providerList,
+  receiverList,
   unitPages
 }) => {
   const [eventVariables, setEventVariables] = React.useState({
@@ -65,35 +48,29 @@ export const EconomicEventManager: React.FC<EconomicEventManagerProps> = ({
     hasNumericalValue: ''
   });
 
-  const [providerLst, setProviderLst] = React.useState<any>([]);
-  const [receiverLst, setReceiverLst] = React.useState<any>([]);
+  const [providerArr, setProviderArr] = React.useState<any>([]);
+  const [receiverArr, setReceiverArr] = React.useState<any>([]);
   const [unitLst, setUnitLst] = React.useState<any>([]);
-  // const [searchValue, setSearchValue] = React.useState<any>({
-  //   action: '',
-  //   provider: '',
-  //   receiver: '',
-  //   hasUnit: '',
-  // })
 
   React.useEffect(() => {
-    if (economicEvents.data?.length) {
-      const providers = economicEvents.data.map((el: any) => ({
-        id: el.provider.id,
-        label: el.provider.name
-      }));
-      const receivers = economicEvents.data.map((el: any) => ({
-        id: el.receiver.id,
-        label: el.receiver.name
-      }));
+    // @ts-ignore
+    const providers = providerList?.map((el: { id: string; name: string }) => ({
+      id: el.id,
+      label: el.name
+    }));
+    // @ts-ignore
+    const receivers = receiverList?.map((el: { id: string; name: string }) => ({
+      id: el.id,
+      label: el.name
+    }));
 
-      setProviderLst(providers);
-      setReceiverLst(receivers);
-    }
-  }, [economicEvents]);
+    setProviderArr(providers);
+    setReceiverArr(receivers);
+  }, [providerList, receiverList]);
 
   React.useEffect(() => {
     if (unitPages?.edges.length) {
-      const unit = unitPages.edges.map((el: any) => ({
+      const unit = unitPages.edges.map((el: { id: string; label: string; symbol: string }) => ({
         id: el.id,
         label: `${el.label} / ${el.symbol}`
       }));
@@ -116,9 +93,12 @@ export const EconomicEventManager: React.FC<EconomicEventManagerProps> = ({
       hasNumericalValue: ''
     });
 
-  const actionHandler = (name: string, option: IntentActions) => {
+  const actionHandler = (name: string, option: IntentActions | null) => {
     eventVariablesHandler({ target: { name, value: option } });
-    setAction(option.id);
+
+    if (option?.id) {
+      setAction(option.id);
+    }
   };
 
   const eventVariablesHandler = (event: any) => {
@@ -128,13 +108,6 @@ export const EconomicEventManager: React.FC<EconomicEventManagerProps> = ({
       [name]: value
     });
   };
-  //
-  // const selectSearchHandler = (name: string, value: string) => {
-  //   if (value.length > 3) {
-  //     console.log(value)
-  //
-  //   }
-  // }
 
   const { create } = useCreateEvent();
 
@@ -172,16 +145,15 @@ export const EconomicEventManager: React.FC<EconomicEventManagerProps> = ({
             id="action"
             name="action"
             onSelect={actionHandler}
-            // onInputChange={selectSearchHandler}
           />
         </FormGroup>
-        {providerLst.length ? (
+        {providerArr.length ? (
           <div>
             <div className="d-flex">
               <FormGroup>
                 <FormLabel>Provider</FormLabel>
                 <Select
-                  options={providerLst}
+                  options={providerArr}
                   variant="primary"
                   value={eventVariables.provider}
                   id="provider"
@@ -193,7 +165,7 @@ export const EconomicEventManager: React.FC<EconomicEventManagerProps> = ({
               <FormGroup>
                 <FormLabel>Receiver</FormLabel>
                 <Select
-                  options={receiverLst}
+                  options={receiverArr}
                   variant="primary"
                   value={eventVariables.receiver}
                   id="receiver"
